requirements.txt
beautifulsoup4==4.12.3
markdownify
matplotlib==3.9.3
networkx==3.3
numpy==2.0.2
gymnasium==0.29.1
playwright==1.49.0
pydantic==2.11.5
python-dotenv==1.0.1
Requests==2.32.3
setuptools
tiktoken==0.7.0
openai==1.99.9
pytest==8.3.4
dependency-injector==4.41.0
flask==3.1.0
lxml==5.3.1
pre-commit==4.1.0
pillow==11.2.1
loguru==0.7.3
rich==13.9.4
xmldiff==2.7.0
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
browser-use==0.7.9
aiohttp==3.12.15
